Oleg rebased last week on main (#2)
* Fixed a bug allowing the launch of multiple overlays.
Fixed an issue where the preview did not become active after taking a screenshot.
Added drag and drop functionality.
The problem with "2+" arises due to the background fill of the overlay.
Do we need to fill the background while taking a screenshot? (The solution depends on this)

* Watched how it works with two monitors (it doesn't); when selecting an area on the second screen, a screenshot of the primary display is taken (learning how to fix this).
Fixed the bug with onHover.
Adjusted the logic of functions.
Hid the application from the dock since we don't need it there.
Window magic is still in progress.
Noticed that the status bar is not displayed, or is it just my issue?

* Fixed bugs with button appearance on the screenshot (previously, the image would disappear)
Fixed window activation
Added a function to open a screenshot in a window; also, created placeholders for functions related to other buttons
Started writing code for the editing window
Window magic is still in progress; changes there are not included in the commit. I'll add them once everything is functioning. (You were right; I was overly optimistic!)

* Fixed a bug with the preview drag-and-drop. Adjusted animations. Tomorrow, I'll commit the changes with the windows, and we'll need to schedule a call to discuss the next steps.

* Now preview stack on the top

---------

Co-authored-by: Oleg <yakushinoa@gmail.com>